requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
bcrypt
python-jose
pypdf
sentence-transformers
faiss-cpu

